Fix: random culled scanlines from pixel center offsets

+++ projects/lab5/code/texture.cc
@@ -94,8 +94,8 @@ void Texture::Unbind()
 Pixel Texture::GetColor(vec3& uvCoord)
 {
 	// fetch position in the texture's buffer, offsetted by bits per pixel
-	const unsigned int x = uvCoord.x * (float)widht;
-	const unsigned int y = uvCoord.y * (float)height;
+	const unsigned int x = uvCoord.x * widht;
+	const unsigned int y = uvCoord.y * height;
 	const unsigned char* imageColor = &localbuf[(x + y * widht) * bpp];
 	
 	return {
